FIX: Fix typo

+++ cjsmem.cpp
@@ -12,7 +12,7 @@ namespace clib {
             data.emplace_back(size);
             return data.back().data();
         }
-        auto&d = data[*removed.begin()];
+        auto &d = data[*removed.begin()];
         removed.erase(removed.begin());
         d.resize(size);
         return d.data();
@@ -23,12 +23,12 @@ namespace clib {
         if (f != map_data.end()) {
             if (data[f->second].size() > 1U << 8U)
                 data[f->second].clear();
-            if (f->second + 1 == data.size()) {
+            if (((size_t) f->second) + 1 == data.size()) {
                 map_data.erase(ptr);
                 data.pop_back();
                 while (!removed.empty()) {
                     auto i = *removed.begin();
-                    if (i + 1 == data.size()) {
+                    if (((size_t) i) + 1 == data.size()) {
                         removed.erase(i);
                         map_data.erase(data.back().data());
                         data.pop_back();
